BarrierSuit works

+++ Stat.h
@@ -4,7 +4,7 @@
 class Stat
 {
 public:
-   virtual int getStat() const = 0;
+   virtual float getStat() const = 0;
    virtual ~Stat() {};
 };
